+++ filter/com_dec.h
@@ -132,7 +132,7 @@ public:
 
   // Stats
   STDMETHODIMP get_frames(int  *frames, int *errors);
-  STDMETHODIMP get_info  (char *info, int len);
+  STDMETHODIMP get_info  (char *info, size_t len);
 
   // Load/save settings
   STDMETHODIMP load_params(Config *config, int what);
@@ -198,8 +198,8 @@ public:
   STDMETHODIMP get_eq_bands     (int *freq, double *gain);
   STDMETHODIMP set_eq_bands     (const int *freq, const double *gain);
   // Spectrum
-  STDMETHODIMP get_spectrum_length (size_t *length);
-  STDMETHODIMP set_spectrum_length (size_t  length);
+  STDMETHODIMP get_spectrum_length (unsigned *length);
+  STDMETHODIMP set_spectrum_length (unsigned  length);
   STDMETHODIMP get_spectrum      (sample_t *data, double *bin2hz);
   // Delay
   STDMETHODIMP get_delay        (bool *delay);
